feat(pricing): a revision number is a name, so nothing gives one back
D-145: the code deleted a discarded draft revision, and the delete freed its
number. `max(revision)` returned to what it was, the next opened draft minted
the same one, and `(plan_id, revision)` — the name the grant table is keyed by,
every revision-scoped child copies under, and the audit trail records — denoted
two different rows over a plan's life. Nothing observed that while the current
revision never regressed, except the entity tag: a client holding the discarded
`plan/2`'s row version patched the new `plan/2` and its precondition passed,
most reachably at the initial version every freshly minted revision carries.
The lost update optimistic concurrency exists to refuse, arriving through the
key instead of the version.

So a discarded draft is now flipped to a terminal `abandoned` state and the row
survives as a tombstone holding its number. `delete_draft` becomes
`abandon_draft` with the compare-and-swap unchanged — one statement, the
submitted version matched inside it, the draft-only predicate beside it — and
the three-way refusal now discriminates one more case than a delete could: a
second abandon is told the revision is already discarded rather than that no
such revision exists. Minting moves to `max(revision) + 1` over the plan's own
rows, because deriving it from the current revision is what hands the tombstone's
number out again; the numbers may therefore have gaps, which is the consequence
the decision states plainly.

`abandoned` falls outside both partial `UNIQUE` predicates, so a plan holds any
number of tombstones beside its one open draft and its one current revision —
verified against the schema rather than assumed. `pricing_plan` now refuses
DELETE outright, on drafts too: the guard is physical because the engine is not
the only thing that can reach the table, and a hand-run delete of a draft
reopens the whole defect. Price rows are untouched — a never-published draft
price row stays deletable, which is where `inst-ps-nodelete` leaves it.

`max_hold_granules` widens to `bigint` and loses its bespoke converter; a
granule count is a duration in the row's own granularity, and at `per_second` an
ordinary hold does not fit an `integer`.

The abandon endpoint and the audit record for the flip are owed: this gear has
no authoring REST surface and no audit writer, and neither was invented here.

+++ gears/bss/pricing/pricing/src/infra/storage/entity/price.rs
@@ -84,7 +84,11 @@ pub struct Model {
     pub aggregation_granularity: Option<String>,
     pub tier_aggregation_window: Option<String>,
     pub tier_qualification_window: Option<String>,
-    pub max_hold_granules: Option<i32>,
+    /// How many granules a level fold may carry a stale reading across. A
+    /// `bigint` like every other count on the row: a granule count is a duration
+    /// in the row's own granularity, and at `per_second` an ordinary
+    /// multi-year hold does not fit an `integer`.
+    pub max_hold_granules: Option<i64>,
     /// The **authored** D-45 declaration `{quantity, rolloverPolicy}`, stored as
     /// written. Slice-10-declared and Slice-10-compiled; this gear persists and
     /// returns it and nothing more, because the D-129 supersession guard reads
